Move dense subgraph resizing functionality to dense
